requirements.txt
absl-py==1.4.0
appdirs==1.4.4
brotlipy==0.7.0
contourpy==1.0.7
cryptography==40.0.2
cycler==0.11.0
decorator==5.1.1
fbm==0.3.0
filelock==3.12.0
fonttools==4.39.3
importlib-resources==5.12.0
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
psutil==5.9.5
pyparsing==3.0.9
python-dateutil==2.8.2
python-telegram-bot==12.8
pytz==2023.3
scipy==1.10.1
scikit-learn==1.2.2
six==1.16.0
sklearn==0.0.post4
sympy==1.11.1
torch==2.0.0
tornado==6.3.1
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
zipp==3.15.0
